pyproject.toml
[project]
name = "rag-openshift-doco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.29.0",
    "chromadb>=0.5.0",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "flask>=3.1.3",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langchain-chroma>=1.1.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ingest", "src/qa", "src/viewer"]
